% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_split_matrix.R
\name{cluster_split_matrix}
\alias{cluster_split_matrix}
\title{Analyzing across cluster solutions}
\usage{
cluster_split_matrix(data, ...)
}
\arguments{
\item{data}{data frame with cluster solutions}

\item{...}{segment variable names, unquoted. tidyselect methods can be used.}
}
\value{
A table showing the split between 2 solutions
}
\description{
\code{cluster_split_matrix} returns a matrix breaking down the split betwween 2 given solutions.
}
\examples{

set.seed(1)
segs_df <- data.frame(
  sol_2 = round(runif(100, min = 1, max = 2), 0),
  sol_3 = round(runif(100, min = 1, max = 3), 0),
  sol_4 = round(runif(100, min = 1, max = 4), 0)
)

cluster_split_matrix(segs_df, sol_2, sol_3)
}
