% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cluster_plot.R
\name{create_cluster_plot}
\alias{create_cluster_plot}
\title{Create Cluster Scatter plot}
\usage{
create_cluster_plot(data, seg, ThreeD = FALSE)
}
\arguments{
\item{data}{The data frame including only the variables used for the segmentation, as well as the segment variable.}

\item{seg}{The name of the segmentation variable in the data frame, unquoted.}

\item{ThreeD}{Logical indicating whether the plot should be 3D or 2D.}
}
\value{
A scatterplot showing the location of the segments across the data space.
}
\description{
Create Cluster Scatter plot
}
\examples{
create_cluster_plot(iris, Species)
create_cluster_plot(iris, Species, ThreeD = TRUE)
}
