% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cluster_tables.R
\name{create_cluster_tables}
\alias{create_cluster_tables}
\title{Create Banner Tables}
\usage{
create_cluster_tables(data, seg, labels, ...)
}
\arguments{
\item{data}{A data frame with the segment variable and the variables to be analyzed}

\item{seg}{The name of the segment variable, unquoted}

\item{labels}{A vector of the labels for the variables being analyzed}

\item{...}{Variables to analyze}
}
\value{
An excel file with banner tables results
}
\description{
Creates a banner table un xlsx format to then analyze outside of R how the segments differ across the variables selected
}
\examples{
\dontrun{
create_cluster_tables(iris, Species, labels = c("Petal Length", "Sepal Length"),
  Petal.Length, Sepal.Length)
}
}
