% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_final_pca.R
\name{create_final_pca}
\alias{create_final_pca}
\title{Finalized Principal Components Table}
\usage{
create_final_pca(data, nfactors, labels_df)
}
\arguments{
\item{data}{A data frame with only the variables to run the PCA on.}

\item{nfactors}{The number of factors desired.}

\item{labels_df}{A data frame that lists the variable names in a "name"
variable and the labels of the those variable in a "label" variable.}
}
\value{
A data frame with the variables ordered by their corresponding factors
}
\description{
Runs a PCA with a pre-specified number of factors and returns a table with
the variables ordered by their factors.
}
\examples{
data <- as.data.frame(Harman74.cor$cov)

labels_df <- data.frame(
  name = names(data),
  label = gsub("([[:lower:]][[:lower:]])([[:upper:]])", "\\\\1 \\\\2", names(data)))

suppressWarnings(
  create_final_pca(data = data, nfactors = 4, labels_df = labels_df)
  )
}
