% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_solutions.R
\name{plot_cluster_solutions}
\alias{plot_cluster_solutions}
\title{Analyzing across cluster solutions}
\usage{
plot_cluster_solutions(data, ...)
}
\arguments{
\item{data}{data frame with cluster solutions}

\item{...}{segment variable names, unquoted.}
}
\value{
An alluvial plot made using ggplot2
}
\description{
\code{plot_cluster_solutions} returns an alluvial plot displaying how segments are broken out from one cluster solution to another.
}
\examples{

set.seed(1)
segs_df <- data.frame(
  sol_2 = round(runif(100, min = 1, max = 2), 0),
  sol_3 = round(runif(100, min = 1, max = 3), 0),
  sol_4 = round(runif(100, min = 1, max = 4), 0)
)

plot_cluster_solutions(segs_df, sol_2, sol_3, sol_4)
}
