% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree_plot.R
\name{scree_plot}
\alias{scree_plot}
\title{Scree plot}
\usage{
scree_plot(data, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{variables to run PCA on. tidyselect methods can be used.}
}
\value{
A scree plot in ggplot2
}
\description{
A scree plot made in ggplot2
}
\examples{

scree_plot(iris, -Species)
}
